+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # CHANGELOG
 
-## Unreleased
-
 ## Todo
 
 - Update the apps index route to display all current features.
+
+## Unreleased
+
+## v0.0.3
+
+**Features**
+
+- Buy X Get Y Product Discount
+  - This function creates a simple (really basic) Buy X Get Y product discount
+  - The function takes 3 arguments: Buy X (number of products), Get Y (number of products), and Percentage (discount percentage)
+  - The function will handle the rest of the logic. Including the discount split. Eligible products depend on a variant metafield: `variant.metafields.debut.enable_b2g1f` \*\* metafield namespace and key could change
+    - Ex: Buy 2 Get 1 Free. Input = 5 eligible products, Discount Split = 4 Paid, 1 Free.
+    - Ex: Buy 2 Get 1 Free. Input = 6 eligible products, Discount Split = 4 Paid, 2 Free.
 
 ## Released